refactor: component panel
