+++ src/components/form-editor/component-settings/datasource-setting.tsx
@@ -3,7 +3,7 @@ import { IFormControl } from '../../../interfaces/form-control';
 import { FormControl, InputLabel, MenuItem, Select, SelectChangeEvent } from '@mui/material';
 import { IParameter } from '../../../interfaces/parameter';
 import { datasourceMatch, getTypesByParameter } from '../../../utils/datasource-match';
-import { generateStandardElement } from '../../../utils/generate-form';
+import {generateElement} from "../../../utils/element-generators";
 
 interface ElementProps {
   value: IFormControl;
@@ -33,7 +33,7 @@ export const DatasourceSetting: React.FC<ElementProps> = (props) => {
     const index = Number(event.target.value);
     const parameter = properties[index];
     if (!datasourceMatch(parameter.type, value.type)) {
-      const standardElement = generateStandardElement(parameter);
+      const standardElement = generateElement(parameter);
       value.properties = standardElement.properties;
       value.type = standardElement.type;
     }
