+++ src/components/form-editor/components-tab/index.tsx
@@ -1,46 +1,18 @@
 import React from 'react';
 import './style.css';
 import { Container, Draggable } from 'react-smooth-dnd';
-import { FORM_GROUP_DIRECTION, IForm } from '../../../interfaces/form-config';
+import { IForm } from '../../../interfaces/form-config';
 import AddCircleOutlineRoundedIcon from '@mui/icons-material/AddCircleOutlineRounded';
-import {CONTROL_TYPE} from "../../../interfaces/form-control";
+import { generateElement, generateGroup, generatePage, generatePageGroup } from '../../../utils/element-generators';
+import { IParameter } from '../../../interfaces/parameter';
 
 interface ComponentsTabProps {
   form: IForm;
-  update: () => void;
+  parameters: IParameter[];
 }
 
 export const ComponentsTab: React.FC<ComponentsTabProps> = (props) => {
-  const { form, update } = props;
-
-  const list: any[] = [
-    {
-      code: 'TabGroup',
-      name: 'Группа табы',
-      pages: [],
-    },
-    {
-      code: 'Page1',
-      name: 'Страница',
-      items: [],
-    },
-    {
-      code: 'Group1',
-      name: 'Группа',
-      direction: FORM_GROUP_DIRECTION.FORCE_HORIZONTAL,
-      items: [],
-    },
-    {
-      code: 'Element1',
-      name: 'Элемент',
-      dataSource: 'Parameter1',
-      type: CONTROL_TYPE.NUMBER,
-      properties: {
-        minValue: 0,
-        maxValue: 100,
-      },
-    },
-  ];
+  const { form, parameters } = props;
 
   const onClick = () => {
     console.log('onClick');
@@ -48,7 +20,7 @@ export const ComponentsTab: React.FC<ComponentsTabProps> = (props) => {
 
   return (
     <div className="tab-item">
-      <Container getChildPayload={() => list[0]} groupName={'pages-groups'} behaviour={'copy'}>
+      <Container getChildPayload={generatePageGroup} groupName={'pages-groups'} behaviour={'copy'}>
         <Draggable>
           <div className="component-item">
             <AddCircleOutlineRoundedIcon style={{ color: '#bcbcd0' }} fontSize={'small'} onClick={onClick} />
@@ -56,7 +28,7 @@ export const ComponentsTab: React.FC<ComponentsTabProps> = (props) => {
           </div>
         </Draggable>
       </Container>
-      <Container getChildPayload={() => list[1]} groupName={'pages'} behaviour={'copy'}>
+      <Container getChildPayload={generatePage} groupName={'pages'} behaviour={'copy'}>
         <Draggable>
           <div className="component-item">
             <AddCircleOutlineRoundedIcon style={{ color: '#bcbcd0' }} fontSize={'small'} onClick={onClick} />
@@ -64,7 +36,7 @@ export const ComponentsTab: React.FC<ComponentsTabProps> = (props) => {
           </div>
         </Draggable>
       </Container>
-      <Container getChildPayload={() => list[2]} groupName={'groups'} behaviour={'copy'}>
+      <Container getChildPayload={generateGroup} groupName={'groups'} behaviour={'copy'}>
         <Draggable>
           <div className="component-item">
             <AddCircleOutlineRoundedIcon style={{ color: '#bcbcd0' }} fontSize={'small'} onClick={onClick} />
@@ -72,14 +44,16 @@ export const ComponentsTab: React.FC<ComponentsTabProps> = (props) => {
           </div>
         </Draggable>
       </Container>
-      <Container getChildPayload={() => list[3]} groupName={'parameters'} behaviour={'copy'}>
-        <Draggable>
-          <div className="component-item">
-            <AddCircleOutlineRoundedIcon style={{ color: '#bcbcd0' }} fontSize={'small'} onClick={onClick} />
-            Поле ввода
-          </div>
-        </Draggable>
-      </Container>
+      {parameters?.length > 0 && (
+        <Container getChildPayload={() => generateElement(parameters[0])} groupName={'parameters'} behaviour={'copy'}>
+          <Draggable>
+            <div className="component-item">
+              <AddCircleOutlineRoundedIcon style={{ color: '#bcbcd0' }} fontSize={'small'} onClick={onClick} />
+              Поле ввода
+            </div>
+          </Draggable>
+        </Container>
+      )}
     </div>
   );
 };
