+++ src/utils/form-generator.ts
@@ -1,0 +1,36 @@
+import { FORM_GROUP_DIRECTION, IForm, IFormItem } from '../interfaces/form-config';
+import { IParameter } from '../interfaces/parameter';
+import { generateElement } from './element-generators';
+
+/** Генерация стандартной формы. */
+export const generateStandardForm = (parameters: IParameter[]): IForm => {
+  const items: IFormItem[] = [];
+  parameters.forEach((param) => {
+    items.push(generateElement(param));
+  });
+  return {
+    code: 'FormCode',
+    name: 'Пример формы',
+    description: 'FormDescription',
+    items: [
+      {
+        code: 'TabGroup',
+        name: 'Группа табы',
+        pages: [
+          {
+            code: 'Page1',
+            name: 'Страница 1',
+            items: [
+              {
+                code: 'Group1',
+                name: 'Группа 1',
+                direction: FORM_GROUP_DIRECTION.FORCE_HORIZONTAL,
+                items: items,
+              },
+            ],
+          },
+        ],
+      },
+    ],
+  };
+};
