+++ src/utils/id-generator.ts
@@ -1,0 +1,5 @@
+let id = 0;
+
+export const generateCode = (): string => {
+  return String(id++);
+};
